refactor : fix incorrect directory name
