+++ src/main/kotlin/com/project/scrumbleserver/controller/handler/ApiControllerAdvice.kt
@@ -1,8 +1,8 @@
 package com.project.scrumbleserver.controller.handler
 
 import com.project.scrumbleserver.global.api.ApiResponse
-import com.project.scrumbleserver.global.excception.BusinessException
-import com.project.scrumbleserver.global.excception.ServerException
+import com.project.scrumbleserver.global.exception.BusinessException
+import com.project.scrumbleserver.global.exception.ServerException
 import mu.KotlinLogging
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -16,7 +16,7 @@ private val logger = KotlinLogging.logger {}
 class ApiControllerAdvice {
 
     data class ErrorResponse(
-        val message: String
+        val message: String,
     )
 
     @ExceptionHandler(BusinessException::class)
